chore: isolate paths members which require fs
